ml/requirements.txt
pandas
scikit-learn
google-cloud-bigquery
google-cloud-bigquery-storage
pyarrow
pandas-gbq